data-challenge/API/CF/requirements.txt
google-auth>=1.31.1
google-cloud-bigquery-storage>=2.4.0
google-cloud-bigquery>=2.20.0
pyarrow>=4.0.0